requirements.txt
cffi==1.15.1
docstring-parser==0.14.1
msgpack==1.0.4
py-algorand-sdk==1.20.2
pycparser==2.21
pycryptodomex==3.16.0
PyNaCl==1.5.0
pyteal==0.20.1
python-dotenv==0.21.0
semantic-version==2.10.0
